Backend/requirements.txt
numpy>=1.26.0
scipy>=1.11.0
flask>=2.3.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
python-socketio>=5.8.0
psutil>=5.9.0
requests>=2.31.0
pymongo>=4.5.0
python-dotenv>=1.0.0
PyJWT>=2.8.0
bcrypt>=4.1.0
